Deleting old image

+++ app.py
@@ -54,7 +54,8 @@ def create_graph(percents):
         showlegend=False
     )
     fig.update_polars(radialaxis_showticklabels=False, radialaxis_showline=False, radialaxis_range=[0, 100])
-    os.remove('static/stats_plot.png')
+    if os.path.exists('static/stats_plot.png'):
+        os.remove('static/stats_plot.png')
     fig.write_image('static/stats_plot.png')
